+++ app/src/main/java/com/concordium/wallet/data/model/TransferSubmissionStatus.kt
@@ -1,12 +1,14 @@
 package com.concordium.wallet.data.model
 
+import java.math.BigInteger
+
 data class TransferSubmissionStatus(
     val status: TransactionStatus,
     val outcome: TransactionOutcome?,
-    val amount: Long?,
+    val amount: BigInteger?,
     val sender: String?,
     val to: String?,
-    val cost: Long?,
+    val cost: BigInteger?,
     val transactionHash: String?,
     val blockHashes: List<String>?,
     val rejectReason: String?,
