+++ app/src/main/java/com/concordium/wallet/ui/account/accountsoverview/AccountsOverviewFragment.kt
@@ -31,8 +31,10 @@ import com.concordium.wallet.ui.common.delegates.IdentityStatusDelegateImpl
 import com.concordium.wallet.ui.identity.identitiesoverview.IdentitiesOverviewActivity
 import com.concordium.wallet.ui.identity.identityproviderlist.IdentityProviderListActivity
 import com.concordium.wallet.ui.transaction.sendfunds.SendFundsActivity
+import java.math.BigInteger
 
-class AccountsOverviewFragment : BaseFragment(), IdentityStatusDelegate by IdentityStatusDelegateImpl(), EarnDelegate by EarnDelegateImpl() {
+class AccountsOverviewFragment : BaseFragment(),
+    IdentityStatusDelegate by IdentityStatusDelegateImpl(), EarnDelegate by EarnDelegateImpl() {
     private var _binding: FragmentAccountsOverviewBinding? = null
     private val binding get() = _binding!!
 
@@ -175,15 +177,32 @@ class AccountsOverviewFragment : BaseFragment(), IdentityStatusDelegate by Ident
             checkAppSettings(appSettings)
         }
         viewModel.localTransfersLoaded.observe(this) { account ->
-            activity?.let { gotoEarn(it, account, viewModel.hasPendingDelegationTransactions, viewModel.hasPendingBakingTransactions) }
+            activity?.let {
+                gotoEarn(
+                    it,
+                    account,
+                    viewModel.hasPendingDelegationTransactions,
+                    viewModel.hasPendingBakingTransactions
+                )
+            }
         }
     }
 
     private fun checkAppSettings(appSettings: AppSettings?) {
         appSettings?.let {
             when (appSettings.status) {
-                AppSettings.APP_VERSION_STATUS_WARNING -> it.url?.let { url -> showAppUpdateWarning(url) }
-                AppSettings.APP_VERSION_STATUS_NEEDS_UPDATE -> it.url?.let { url -> showAppUpdateNeedsUpdate(url) }
+                AppSettings.APP_VERSION_STATUS_WARNING -> it.url?.let { url ->
+                    showAppUpdateWarning(
+                        url
+                    )
+                }
+
+                AppSettings.APP_VERSION_STATUS_NEEDS_UPDATE -> it.url?.let { url ->
+                    showAppUpdateNeedsUpdate(
+                        url
+                    )
+                }
+
                 else -> {}
             }
         }
@@ -233,7 +252,8 @@ class AccountsOverviewFragment : BaseFragment(), IdentityStatusDelegate by Ident
         if (App.appCore.closingPoolsChecked)
             return
 
-        App.appCore.closingPoolsChecked = true // avoid calling this more than once for each app cold launch
+        App.appCore.closingPoolsChecked =
+            true // avoid calling this more than once for each app cold launch
 
         val poolIds = accountList.mapNotNull { accountWithIdentity ->
             accountWithIdentity.account.accountDelegation?.delegationTarget?.bakerId?.toString()
@@ -256,8 +276,11 @@ class AccountsOverviewFragment : BaseFragment(), IdentityStatusDelegate by Ident
                 if (affectedAccounts.isNotEmpty()) {
                     val builder = AlertDialog.Builder(context)
                     builder.setTitle(R.string.accounts_overview_closing_pools_notice_title)
-                    builder.setMessage(getString(R.string.accounts_overview_closing_pools_notice_message).plus(
-                        affectedAccounts))
+                    builder.setMessage(
+                        getString(R.string.accounts_overview_closing_pools_notice_message).plus(
+                            affectedAccounts
+                        )
+                    )
                     builder.setPositiveButton(getString(R.string.accounts_overview_closing_pools_notice_ok)) { dialog, _ ->
                         dialog.dismiss()
                     }
@@ -272,17 +295,29 @@ class AccountsOverviewFragment : BaseFragment(), IdentityStatusDelegate by Ident
     private fun checkForUnencrypted(accountList: List<AccountWithIdentity>) {
         accountList.forEach {
 
-            val hasUnencryptedTransactions = it.account.finalizedEncryptedBalance?.incomingAmounts?.isNotEmpty()
-            if((hasUnencryptedTransactions != null && hasUnencryptedTransactions == true)
+            val hasUnencryptedTransactions =
+                it.account.finalizedEncryptedBalance?.incomingAmounts?.isNotEmpty()
+            if ((hasUnencryptedTransactions != null && hasUnencryptedTransactions == true)
                 && it.account.transactionStatus == TransactionStatus.FINALIZED
                 && !App.appCore.session.isShieldedWarningDismissed(it.account.address)
                 && !App.appCore.session.isShieldingEnabled(it.account.address)
-                && encryptedWarningDialog == null){
+                && encryptedWarningDialog == null
+            ) {
 
                 val builder = AlertDialog.Builder(context)
                 builder.setTitle(getString(R.string.account_details_shielded_warning_title))
-                builder.setMessage(getString(R.string.account_details_shielded_warning_text, it.account.name))
-                builder.setNegativeButton(getString(R.string.account_details_shielded_warning_enable, it.account.name)) { _, _ ->
+                builder.setMessage(
+                    getString(
+                        R.string.account_details_shielded_warning_text,
+                        it.account.name
+                    )
+                )
+                builder.setNegativeButton(
+                    getString(
+                        R.string.account_details_shielded_warning_enable,
+                        it.account.name
+                    )
+                ) { _, _ ->
                     startShieldedIntroFlow(it.account)
                     encryptedWarningDialog?.dismiss()
                     encryptedWarningDialog = null
@@ -328,7 +363,7 @@ class AccountsOverviewFragment : BaseFragment(), IdentityStatusDelegate by Ident
 
         eventListener = object : Preferences.Listener {
             override fun onChange() {
-                if(isInLayout && isVisible && !isDetached){
+                if (isInLayout && isVisible && !isDetached) {
                     updateWarnings()
                 }
             }
@@ -337,12 +372,13 @@ class AccountsOverviewFragment : BaseFragment(), IdentityStatusDelegate by Ident
         initializeList()
     }
 
-    private fun updateWarnings(){
+    private fun updateWarnings() {
         val identity = viewModel.pendingIdentityForWarningLiveData.value
         if (identity != null && !App.appCore.session.isIdentityPendingWarningAcknowledged(identity.id)) {
             binding.identityPending.visibility = View.VISIBLE
             viewModel.pendingIdentityForWarningLiveData.value
-            binding.identityPendingTv.text = getString(R.string.accounts_overview_identity_pending_warning, identity.name)
+            binding.identityPendingTv.text =
+                getString(R.string.accounts_overview_identity_pending_warning, identity.name)
             binding.identityPendingClose.setOnClickListener {
                 App.appCore.session.setIdentityPendingWarningAcknowledged(identity.id)
                 updateWarnings()
@@ -350,8 +386,7 @@ class AccountsOverviewFragment : BaseFragment(), IdentityStatusDelegate by Ident
             binding.identityPending.setOnClickListener {
                 startActivity(Intent(activity, IdentitiesOverviewActivity::class.java))
             }
-        }
-        else {
+        } else {
             binding.identityPending.visibility = View.GONE
         }
     }
@@ -452,13 +487,15 @@ class AccountsOverviewFragment : BaseFragment(), IdentityStatusDelegate by Ident
         binding.createAccountButton.visibility = state
     }
 
-    private fun showTotalBalance(totalBalance: Long) {
+    private fun showTotalBalance(totalBalance: BigInteger) {
         binding.totalBalanceTextview.text = CurrencyUtil.formatGTU(totalBalance)
     }
-    private fun showDisposalBalance(atDisposal: Long) {
+
+    private fun showDisposalBalance(atDisposal: BigInteger) {
         binding.accountsOverviewTotalDetailsDisposal.text = CurrencyUtil.formatGTU(atDisposal, true)
     }
-    private fun showStakedBalance(totalBalance: Long) {
+
+    private fun showStakedBalance(totalBalance: BigInteger) {
         binding.accountsOverviewTotalDetailsStaked.text = CurrencyUtil.formatGTU(totalBalance, true)
     }
